+++ changelog.md
@@ -1,9 +1,11 @@
 ## OCHRE Changelog
 
 ### Updates from PRs
-
 - Added startup capacity degredation for ASHP/MSHP/AC
   [#179](https://github.com/NREL/OCHRE/issues/179)
+- Fixed bug in EV equivalent battery model (EBM) results
+- Updated `Analysis.combine_time_series_files` to aggregate across dwellings
+  only
 
 ### OCHRE v0.9.1